Resources (#1)
* Add AnimationController and SkeletonPart

* Add AnimationProperties and Enum array serialization support.

* Update CPU target to x86

* Update serialization and add resources

* Add Collision

* Add CollisionMesh

* Update float datatypes for proper render

* Change FileRef to AsciiString for binary file

* Add geometry

* Fix memory offset

* Fix serialization type

* Add ModelMorphSettings

* Add particle animation

* Add skeletal animation

* Add Widget Layer

* Add Texture and UITexture

* Add TextureAtlas

* Add TexturePatch

* Add UiTextureItem

* Code cleanup

* Fix serialization and add log

* Add exception details

* Update AsciiString array

* Fix AsciiString reading

* Add TextFileRef type

* Add WidgetButton

* Add WidgetPanel

* Add WidgetTextView

* Add TextureSingleElement

* Add MobAnimations

* Fix missing Resource class extension

* Add generic field

* Code cleanup

* Improve AsciiString serialization

* Remove TODO

* Add log to GenericField

* Add VisualMob

* Add TextureElementType

* Add VisObjectTemplate

* Code cleanup

* Quick test

+++ Database/DataType/Implementation/Quaternion.cs
@@ -21,10 +21,10 @@ namespace Database.DataType.Implementation
         public override XElement Serialize(string name)
         {
             return new XElement(name,
-                new XAttribute("x", _x),
-                new XAttribute("y", _y),
-                new XAttribute("z", _z),
-                new XAttribute("w", _w));
+                new XAttribute("x", (float) _x),
+                new XAttribute("y", (float) _y),
+                new XAttribute("z", (float) _z),
+                new XAttribute("w", (float) _w));
         }
 
         public override void Deserialize(IntPtr memoryAddress)
